+++ fixtures/app_with_development_region_config/Config.swift
@@ -1,7 +1,7 @@
 import ProjectDescription
 
-let config = Config(
+let config = TuistConfig(
   generationOptions: [
-    .organizationName("Tuist")
+    .developmentRegion("de")
   ]
 )